Add test 에서 FromContainer test 분리
